Resolved saving issue and loading in Pokemon into Box
Defined nullable pokemon values, and detection within Pokemon class is parameter HasValue. Parameter checks if pokemon is connected to a scriptable object, if so it is set to true, otherwise it is set to false.

Issue with saving related to Box background image, as Image is not naturally savable in binary format.
Removal of box background image from saving methods resolved issue.